+++ minos/common/model/aggregate.py
@@ -17,6 +17,10 @@ from typing import (
     Optional,
 )
 
+from dependency_injector.wiring import (
+    Provide,
+)
+
 from ..configuration import (
     MinosConfig,
 )
@@ -29,6 +33,9 @@ from ..exceptions import (
 )
 from ..meta import (
     self_or_classmethod,
+)
+from ..networks import (
+    MinosBroker,
 )
 from ..repository import (
     MinosRepository,
@@ -45,19 +52,32 @@ class Aggregate(MinosModel):
     id: int
     version: int
 
-    # FIXME: The ``broker`` attribute should be a reference to a ``MinosBaseBroker`` class instance.
-    # noinspection PyShadowingBuiltins
+    _broker: MinosBroker = Provide["event_broker"]
+    _repository: MinosRepository = Provide["repository"]
+
     def __init__(
-        self, id: int, version: int, *args, _broker: str = None, _repository: MinosRepository = None, **kwargs,
+        self,
+        id: int,
+        version: int,
+        *args,
+        _broker: Optional[MinosBroker] = None,
+        _repository: Optional[MinosRepository] = None,
+        **kwargs,
     ):
 
         super().__init__(id, version, *args, **kwargs)
-        self._broker = _broker
-        self._repository = _repository
+        if _broker is not None:
+            self._broker = _broker
+        if _repository is not None:
+            self._repository = _repository
 
     @classmethod
     async def get(
-        cls, ids: list[int], _config: MinosConfig = None, _broker: str = None, _repository: MinosRepository = None
+        cls,
+        ids: list[int],
+        _config: Optional[MinosConfig] = None,
+        _broker: Optional[MinosBroker] = None,
+        _repository: Optional[MinosRepository] = None,
     ) -> list[Aggregate]:
         """Get a sequence of aggregates based on a list of identifiers.
 
@@ -73,7 +93,11 @@ class Aggregate(MinosModel):
     # noinspection PyShadowingBuiltins
     @classmethod
     async def get_one(
-        cls, id: int, _config: MinosConfig = None, _broker: str = None, _repository: MinosRepository = None
+        cls,
+        id: int,
+        _config: Optional[MinosConfig] = None,
+        _broker: Optional[MinosBroker] = None,
+        _repository: Optional[MinosRepository] = None,
     ) -> Aggregate:
         """Get one aggregate based on an identifier.
 
@@ -85,7 +109,9 @@ class Aggregate(MinosModel):
         :return: An aggregate instance.
         """
         if _repository is None:
-            _repository = cls._build_repository(_config)
+            _repository = cls._repository
+        if _repository is None or isinstance(_repository, Provide):
+            raise MinosRepositoryNonProvidedException("A repository instance is required.")
 
         # noinspection PyTypeChecker
         entries = [v async for v in _repository.select(aggregate_name=cls.classname, aggregate_id=id)]
@@ -103,7 +129,12 @@ class Aggregate(MinosModel):
 
     @classmethod
     async def create(
-        cls, *args, _config: MinosConfig = None, _broker: str = None, _repository: MinosRepository = None, **kwargs
+        cls,
+        *args,
+        _config: Optional[MinosConfig] = None,
+        _broker: Optional[MinosBroker] = None,
+        _repository: Optional[MinosRepository] = None,
+        **kwargs,
     ) -> Aggregate:
         """Create a new ``Aggregate`` instance.
 
@@ -125,10 +156,12 @@ class Aggregate(MinosModel):
             )
 
         if _broker is None:
-            _broker = "MinosBaseBroker()"
+            _broker = cls._broker
 
         if _repository is None:
-            _repository = cls._build_repository(_config)
+            _repository = cls._repository
+            if isinstance(_repository, Provide):
+                raise MinosRepositoryNonProvidedException("A repository instance is required.")
 
         instance = cls(0, 0, *args, _broker=_broker, _repository=_repository, **kwargs)
 
@@ -142,13 +175,19 @@ class Aggregate(MinosModel):
     # noinspection PyMethodParameters,PyShadowingBuiltins
     @self_or_classmethod
     async def update(
-        self_or_cls, id: int = None, _config: MinosConfig = None, _repository: MinosRepository = None, **kwargs
+        self_or_cls,
+        id: Optional[int] = None,
+        _config: Optional[MinosConfig] = None,
+        _broker: Optional[MinosBroker] = None,
+        _repository: Optional[MinosRepository] = None,
+        **kwargs,
     ) -> Aggregate:
         """Update an existing ``Aggregate`` instance.
 
         :param id: If the method call is performed from an instance the identifier is ignored, otherwise it is used to
             identify the target instance.
         :param _config: Current minos config.
+        :param _broker: Broker to be set to the aggregates.
         :param _repository: Repository to be set to the aggregate.
         :param kwargs: Additional named arguments.
         :return: An updated ``Aggregate``  instance.
@@ -158,8 +197,13 @@ class Aggregate(MinosModel):
                 f"The version must be computed internally on the repository. Obtained: {kwargs['version']}"
             )
 
+        if _broker is None:
+            _broker = self_or_cls._broker
+
         if _repository is None:
-            _repository = self_or_cls._build_repository(_config)
+            _repository = self_or_cls._repository
+            if isinstance(_repository, Provide):
+                raise MinosRepositoryNonProvidedException("A repository instance is required.")
 
         if isinstance(self_or_cls, type):
             assert issubclass(self_or_cls, Aggregate)
@@ -187,18 +231,28 @@ class Aggregate(MinosModel):
     # noinspection PyMethodParameters,PyShadowingBuiltins
     @self_or_classmethod
     async def delete(
-        self_or_cls, id: Optional[int] = None, _config: MinosConfig = None, _repository: MinosRepository = None
+        self_or_cls,
+        id: Optional[int] = None,
+        _config: Optional[MinosConfig] = None,
+        _broker: Optional[MinosBroker] = None,
+        _repository: Optional[MinosRepository] = None,
     ):
         """Delete the given aggregate instance.
 
         :param id: If the method call is performed from an instance the identifier is ignored, otherwise it is used to
             identify the target instance.
         :param _config: Current minos config.
+        :param _broker: Broker to be set to the aggregates.
         :param _repository: Repository to be set to the aggregate.
         :return: This method does not return anything.
         """
+        if _broker is None:
+            _broker = self_or_cls._broker
+
         if _repository is None:
-            _repository = self_or_cls._build_repository(_config)
+            _repository = self_or_cls._repository
+            if isinstance(_repository, Provide):
+                raise MinosRepositoryNonProvidedException("A repository instance is required.")
 
         if isinstance(self_or_cls, type):
             assert issubclass(self_or_cls, Aggregate)
@@ -207,24 +261,3 @@ class Aggregate(MinosModel):
             instance = self_or_cls
 
         await _repository.delete(instance)
-
-    # noinspection PyMethodParameters
-    @self_or_classmethod
-    def _build_repository(self_or_cls, config: MinosConfig) -> MinosRepository:
-        repository = None
-
-        if not isinstance(self_or_cls, type):
-            repository = self_or_cls._repository
-
-        if repository is None:
-            # FIXME: In the future, this could be parameterized.
-            from ..repository import (
-                PostgreSqlMinosRepository,
-            )
-
-            repository = PostgreSqlMinosRepository.from_config(already_setup=True, config=config)
-
-        if repository is None:
-            raise MinosRepositoryNonProvidedException("A repository instance is required.")
-
-        return repository