+++ tests/test_common/test_model/test_aggregate.py
@@ -5,19 +5,18 @@ This file is part of minos framework.
 
 Minos framework can not be copied and/or distributed without the express permission of Clariteia SL.
 """
+import sys
 import unittest
 
 from minos.common import (
+    MinosConfig,
+    MinosDependencyInjector,
     MinosInMemoryRepository,
     MinosRepositoryAggregateNotFoundException,
     MinosRepositoryDeletedAggregateException,
     MinosRepositoryManuallySetAggregateIdException,
     MinosRepositoryManuallySetAggregateVersionException,
     MinosRepositoryNonProvidedException,
-    PostgreSqlMinosRepository,
-)
-from minos.common.testing import (
-    PostgresAsyncTestCase,
 )
 from tests.aggregate_classes import (
     Car,
@@ -28,6 +27,17 @@ from tests.utils import (
 
 
 class TestAggregate(unittest.IsolatedAsyncioTestCase):
+    def setUp(self) -> None:
+        self.config_file_path = BASE_PATH / "test_config.yml"
+        self.config = MinosConfig(path=str(self.config_file_path))
+
+    async def asyncSetUp(self):
+        self.injector = MinosDependencyInjector(self.config)
+        await self.injector.wire(modules=[sys.modules[__name__]])
+
+    async def asyncTearDown(self):
+        await self.injector.unwire()
+
     async def test_create(self):
         async with MinosInMemoryRepository() as repository:
             car = await Car.create(doors=3, color="blue", _repository=repository)
@@ -126,36 +136,6 @@ class TestAggregate(unittest.IsolatedAsyncioTestCase):
             await Car.delete(car.id, _repository=repository)
             with self.assertRaises(MinosRepositoryDeletedAggregateException):
                 await Car.get_one(car.id, _repository=repository)
-
-
-class TestAggregateWithConfig(PostgresAsyncTestCase):
-    CONFIG_FILE_PATH = BASE_PATH / "test_config.yml"
-
-    async def test_update(self):
-        with self.config:
-            await PostgreSqlMinosRepository.from_config().setup()
-
-            car = await Car.create(doors=3, color="blue")
-
-            await car.update(color="red")
-            self.assertEqual(Car(1, 2, 3, "red"), car)
-            self.assertEqual(car, await Car.get_one(car.id))
-
-            await car.update(doors=5)
-            self.assertEqual(Car(1, 3, 5, "red"), car)
-            self.assertEqual(car, await Car.get_one(car.id))
-
-            await car.delete()
-            with self.assertRaises(MinosRepositoryDeletedAggregateException):
-                await Car.get_one(car.id)
-
-            car = await Car.create(doors=3, color="blue")
-            await Car.update(car.id, color="red")
-            self.assertEqual(Car(2, 2, 3, "red"), await Car.get_one(car.id))
-
-            await Car.delete(car.id)
-            with self.assertRaises(MinosRepositoryDeletedAggregateException):
-                await Car.get_one(car.id)
 
 
 if __name__ == "__main__":